refactored password and profile pages
