+++ views/forgot-password.ejs
@@ -7,7 +7,7 @@
             <p class="col-lg-10 fs-4">Enter your email address and we'll send you a link to reset your password.</p>
         </div>
         <div class="col-md-10 mx-auto col-lg-5">
-            <form class="p-4 p-md-5 border rounded-3 bg-body-tertiary" action="/forgot-password" method="post">
+            <form class="p-4 p-md-5 border rounded-3 bg-body-tertiary" action="/password/forgot-password" method="post">
                 <div class="form-floating mb-3">
                     <input name="email" type="email" class="form-control" id="floatingInput" placeholder="name@example.com" required>
                     <label for="floatingInput">Email address</label>
