+++ views/login.ejs
@@ -21,7 +21,7 @@
         </form>
         <div class="text-center">
           <br>
-          <a href="/forgot-password">Forgot your password?</a>
+          <a href="/password/forgot-password">Forgot your password?</a>
         </div>
       </div>
     </div>
