+++ views/profile.ejs
@@ -17,7 +17,7 @@
                     </div>
                     <div class="form-group">
                         <label for="dietaryRestrictions">Dietary Restrictions:</label>
-                        <% if (dietaryRestrictions && dietaryRestrictions.length &&dietaryRestrictions[0] !== null) { %>
+                        <% if (dietaryRestrictions && dietaryRestrictions.length && dietaryRestrictions[0] !== null) { %>
                         <ul>
                             <% dietaryRestrictions.forEach(function(item) { %>
                             <li><%= item %></li>
@@ -27,7 +27,7 @@
                         <p>No dietary restrictions specified.</p>
                         <% } %>
                     </div>
-                    <a href="/editProfile" class="btn btn-primary">Edit Profile</a>
+                    <a href="/profile/editProfile" class="btn btn-primary">Edit Profile</a>
                 </div>
             </div>
         </div>
