+++ views/reset-password.ejs
@@ -1,6 +1,6 @@
 <%- include("templates/header") %> 
 <h1>Password Reset</h1>
-<form action="/reset-password" method="post">
+<form action="/password/reset-password" method="post">
     <input type="hidden" name="token" value="<%= token %>">
     <label for="password">New Password:</label>
     <input type="password" id="password" name="password" required>
