+++ views/templates/header.ejs
@@ -22,7 +22,7 @@
     
           <ul class="nav nav-pills">
             <li class="nav-item"><a href="/home" class="nav-link active" aria-current="page">Home</a></li>
-            <li class="nav-item"><a href="/profile" class="nav-link">Profile</a></li>
+            <li class="nav-item"><a href="/profile/" class="nav-link">Profile</a></li>
             <li class="nav-item"><a href="/logout" class="nav-link">Logout</a></li>
           </ul>
         </header>
